This commit syncs constants.js with the python sample API code from IB.
When node-ib used a different name than IB, the old name was kept for
compatibility.

A script was used to programatically compare the old data and the new
data.  The only difference is the min server version for
SEC_DEF_OPT_PARAMS_REQ which was my own error from commit #100